+++ internal/k8s/resource_structs.go
@@ -2,11 +2,11 @@ package k8s
 
 type Resource struct {
 	Name, Namespace           string
-	RequestCPU, RequestMemory float32
-	LimitCPU, LimitMemory     float32
+	RequestCPU, RequestMemory float64
+	LimitCPU, LimitMemory     float64
 }
 
 type ResourceUsage struct {
 	PodName     string
-	CPU, Memory float32
+	CPU, Memory float64
 }
